+++ src/windows/enumerate.rs
@@ -185,9 +185,30 @@ fn parse_usb_port_info(hardware_id: &str, parent_hardware_id: Option<&str>) -> O
         manufacturer: None,
         product: None,
 
+        #[cfg(feature = "usbportinfo-location")]
+        location: None,
+
         #[cfg(feature = "usbportinfo-interface")]
         interface,
     })
+}
+
+#[cfg(feature = "usbportinfo-location")]
+fn parse_location_path(s: &str) -> Option<crate::Location> {
+    let usbroot = "#USBROOT(";
+    let start_i = s.find(usbroot)?;
+    let close_i = s[start_i + usbroot.len()..].find(')')?;
+    let (bus, mut s) = s.split_at(start_i + usbroot.len() + close_i + 1);
+
+    let mut path = Vec::new();
+
+    while let Some((_, next)) = s.split_once("#USB(") {
+        let (port_num, next) = next.split_once(")")?;
+        path.push(port_num.parse().ok()?);
+        s = next;
+    }
+
+    Some(crate::Location::new(bus.to_owned(), path))
 }
 
 struct PortDevices {
@@ -394,6 +415,16 @@ impl PortDevice {
             .map(|mut info: UsbPortInfo| {
                 info.manufacturer = self.property(SPDRP_MFG);
                 info.product = self.property(SPDRP_FRIENDLYNAME);
+
+                #[cfg(feature = "usbportinfo-location")]
+                {
+                    use windows_sys::Win32::Devices::DeviceAndDriverInstallation:: SPDRP_LOCATION_PATHS;
+                    let location_paths = self.property_list(SPDRP_LOCATION_PATHS);
+                    info.location = location_paths
+                        .iter()
+                        .find_map(|p| parse_location_path(p));
+                }
+
                 SerialPortType::UsbPort(info)
             })
             .unwrap_or(SerialPortType::Unknown)
@@ -428,6 +459,41 @@ impl PortDevice {
             .split(';')
             .next_back()
             .map(str::to_string)
+    }
+
+    // Retrieves a device property and returns it, if it exists. Returns None if the property
+    // doesn't exist.
+    #[cfg(feature = "usbportinfo-location")]
+    fn property_list(&mut self, property_id: u32) -> Vec<String> {
+        use windows_sys::Win32::System::Registry::REG_MULTI_SZ;
+        let mut value_type = 0;
+        let mut property_buf = [0u16; MAX_PATH as usize];
+
+        let res = unsafe {
+            SetupDiGetDeviceRegistryPropertyW(
+                self.hdi,
+                &self.devinfo_data,
+                property_id,
+                &mut value_type,
+                property_buf.as_mut_ptr() as *mut u8,
+                property_buf.len() as u32,
+                ptr::null_mut(),
+            )
+        };
+
+        if res == FALSE || value_type != REG_MULTI_SZ {
+            return vec![];
+        }
+
+        let mut out = vec![];
+        for chunk in property_buf.split(|c| *c == 0) {
+            // A NULL value indicates the end of the list
+            if chunk.is_empty() {
+                break;
+            }
+            out.push(from_utf16_lossy_trimmed(chunk));
+        }
+        out
     }
 }