+++ CHANGELOG.md
@@ -8,18 +8,58 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 
 ### Added
+
+* Provide USB port location in terms of bus and port chain via `UsbPortInfo`
+  and `Location`.
+  [#222](https://github.com/serialport/serialport-rs/issues/222)
+  [#350](https://github.com/serialport/serialport-rs/pull/350)
+  [#360](https://github.com/serialport/serialport-rs/pull/360)
+
+### Changed
+### Fixed
+### Removed
+
+
+## [4.9.0] - 2026-03-12
+
+### Added
+
+* Add `exclusive` configuration option to `SerialPortBuilder` to control
+  whether ports are opened with exclusive access locks.
+  [#319](https://github.com/serialport/serialport-rs/pull/319)
+
 ### Changed
 
 * Enumerate serial-base subsystem on Linux without libudev support too. Some of
   these newly enumerated ports may show up with a different type than with
   libudev, but they will be finally enumerated at all.
   [#311](https://github.com/serialport/serialport-rs/pull/311)
+* Enable arbitrary baud rates for Linux musl targets by directly using
+  `termios2` and the corresponding `ioctl`s `TCGETS2` and `TCSETS2` from nix
+  and the libc crate. There is no support form musl for `termios2` and we are
+  directly talking to the kernel and bypassing it here.
+  [#316](https://github.com/serialport/serialport-rs/pull/316)
 
 ### Fixed
-### Removed
+
+* Always propagate errors out of termios getter/setter implementations.
+* Wrap slave port file descriptors in `TTYPort::pair` in a `Drop`-safe `OwnedFd` to avoid file descriptor leaks on early returns.
+* Enable arbitrary baud rates for Linux musl targets by using the `termios2`
+  (`TCGETS2`/`TCSETS2` + `BOTHER`) path now that required `libc` symbols are
+  available.
+  [#316](https://github.com/serialport/serialport-rs/pull/316)
+* Unpin dependencies with MSRV incompatible Rust version bump. This caused
+  issues with dependency resolution for many users building with newer Rust
+  versions. See [Dependencies](README.md#dependencies) on how to pin them in
+  your project when building with older Rust versions.
+  [#324](https://github.com/serialport/serialport-rs/issues/324)
+  [#307](https://github.com/serialport/serialport-rs/issues/307)
+  [#304](https://github.com/serialport/serialport-rs/issues/304)
+  [#300](https://github.com/serialport/serialport-rs/issues/300)
+  [#231](https://github.com/serialport/serialport-rs/issues/231)
 
 
-## [4.8.1] - 2025-10-07i
+## [4.8.1] - 2025-10-07
 
 ### Fixed
 
@@ -59,13 +99,13 @@ project adheres to [Semantic Versioning](https://semver.org/).
   [#251](https://github.com/serialport/serialport-rs/issues/251)
   [#243](https://github.com/serialport/serialport-rs/issues/243)
   [#239](https://github.com/serialport/serialport-rs/pull/239)
-  [#29](https://github.com/serialport/serialport-rs/pull/29)
+  [#29](https://github.com/serialport/serialport-rs/issues/29)
 
 ### Fixed
 
 * Fix reporting serial numbers with colons on Windows.
   [#279](https://github.com/serialport/serialport-rs/issues/279)
-  [#282](https://github.com/serialport/serialport-rs/issues/282)
+  [#282](https://github.com/serialport/serialport-rs/pull/282)
 * Setting arbitrary baud rates on Linux which resulted in issues when read back
   on Arch recently.
   [#281](https://github.com/serialport/serialport-rs/issues/281)
@@ -571,7 +611,8 @@ Unreleased, happened due to a user error using `cargo-release`.
 * Initial release.
 
 
-[Unreleased]: https://github.com/serialport/serialport-rs/compare/v4.8.1...HEAD
+[Unreleased]: https://github.com/serialport/serialport-rs/compare/v4.9.0...HEAD
+[4.9.0]: https://github.com/serialport/serialport-rs/compare/v4.8.1...v4.9.0
 [4.8.1]: https://github.com/serialport/serialport-rs/compare/v4.8.0...v4.8.1
 [4.8.0]: https://github.com/serialport/serialport-rs/compare/v4.7.3...v4.8.0
 [4.7.3]: https://github.com/serialport/serialport-rs/compare/v4.7.2...v4.7.3